Bug 1224849 - Extract Authenticator handling from LoginsHelper
