+++ Client/Frontend/Browser/LoginsHelper.swift
@@ -13,14 +13,11 @@ private let log = Logger.browserLogger
 private let NotNowButtonTitle = NSLocalizedString("Not now", comment: "Button to not save the user's password")
 private let UpdateButtonTitle = NSLocalizedString("Update", comment: "Button to update the user's password")
 private let YesButtonTitle = NSLocalizedString("Yes", comment: "Button to save the user's password")
-private let CancelButtonTitle = NSLocalizedString("Cancel", comment: "Authentication prompt cancel button")
-private let LogInButtonTitle  = NSLocalizedString("Log in", comment: "Authentication prompt log in button")
 
 class LoginsHelper: BrowserHelper {
     private weak var browser: Browser?
     private let profile: Profile
     private var snackBar: SnackBar?
-    private static let MaxAuthenticationAttempts = 3
 
     class func name() -> String {
         return "LoginsHelper"
@@ -92,7 +89,11 @@ class LoginsHelper: BrowserHelper {
         return attr
     }
 
-    private func setCredentials(login: LoginData) {
+    func getLoginsForProtectionSpace(protectionSpace: NSURLProtectionSpace) -> Deferred<Maybe<Cursor<LoginData>>> {
+        return profile.logins.getLoginsForProtectionSpace(protectionSpace)
+    }
+
+    func setCredentials(login: LoginData) {
         if login.password.isEmpty {
             log.debug("Empty password")
             return
@@ -205,89 +206,4 @@ class LoginsHelper: BrowserHelper {
             })
         }
     }
-
-    func handleAuthRequest(viewController: UIViewController, challenge: NSURLAuthenticationChallenge) -> Deferred<Maybe<LoginData>> {
-        // If there have already been too many login attempts, we'll just fail.
-        if challenge.previousFailureCount >= LoginsHelper.MaxAuthenticationAttempts {
-            return deferMaybe(LoginDataError(description: "Too many attempts to open site"))
-        }
-
-        var credential = challenge.proposedCredential
-
-        // If we were passed an initial set of credentials from iOS, try and use them.
-        if let proposed = credential {
-            if !(proposed.user?.isEmpty ?? true) {
-                if challenge.previousFailureCount == 0 {
-                    return deferMaybe(Login.createWithCredential(credential!, protectionSpace: challenge.protectionSpace))
-                }
-            } else {
-                credential = nil
-            }
-        }
-
-        if let credential = credential {
-            // If we have some credentials, we'll show a prompt with them.
-            return promptForUsernamePassword(viewController, credentials: credential, protectionSpace: challenge.protectionSpace)
-        }
-
-        // Otherwise, try to look one up
-        return profile.logins.getLoginsForProtectionSpace(challenge.protectionSpace).bindQueue(dispatch_get_main_queue()) { res in
-            let credentials = res.successValue?[0]?.credentials
-            return self.promptForUsernamePassword(viewController, credentials: credentials, protectionSpace: challenge.protectionSpace)
-        }
-    }
-
-    private func promptForUsernamePassword(viewController: UIViewController, credentials: NSURLCredential?, protectionSpace: NSURLProtectionSpace) -> Deferred<Maybe<LoginData>> {
-        if protectionSpace.host.isEmpty {
-            print("Unable to show a password prompt without a hostname")
-            return deferMaybe(LoginDataError(description: "Unable to show a password prompt without a hostname"))
-        }
-
-        let deferred = Deferred<Maybe<LoginData>>()
-        let alert: UIAlertController
-        let title = NSLocalizedString("Authentication required", comment: "Authentication prompt title")
-        if !(protectionSpace.realm?.isEmpty ?? true) {
-            let msg = NSLocalizedString("A username and password are being requested by %@. The site says: %@", comment: "Authentication prompt message with a realm. First parameter is the hostname. Second is the realm string")
-            let formatted = NSString(format: msg, protectionSpace.host, protectionSpace.realm ?? "") as String
-            alert = UIAlertController(title: title, message: formatted, preferredStyle: UIAlertControllerStyle.Alert)
-        } else {
-            let msg = NSLocalizedString("A username and password are being requested by %@.", comment: "Authentication prompt message with no realm. Parameter is the hostname of the site")
-            let formatted = NSString(format: msg, protectionSpace.host) as String
-            alert = UIAlertController(title: title, message: formatted, preferredStyle: UIAlertControllerStyle.Alert)
-        }
-
-        // Add a button to log in.
-        let action = UIAlertAction(title: LogInButtonTitle,
-            style: UIAlertActionStyle.Default) { (action) -> Void in
-                guard let user = alert.textFields?[0].text, pass = alert.textFields?[1].text else { deferred.fill(Maybe(failure: LoginDataError(description: "Username and Password required"))); return }
-
-                let login = Login.createWithCredential(NSURLCredential(user: user, password: pass, persistence: .ForSession), protectionSpace: protectionSpace)
-                deferred.fill(Maybe(success: login))
-                self.setCredentials(login)
-        }
-        alert.addAction(action)
-
-        // Add a cancel button.
-        let cancel = UIAlertAction(title: CancelButtonTitle, style: UIAlertActionStyle.Cancel) { (action) -> Void in
-            deferred.fill(Maybe(failure: LoginDataError(description: "Save password cancelled")))
-        }
-        alert.addAction(cancel)
-
-        // Add a username textfield.
-        alert.addTextFieldWithConfigurationHandler { (textfield) -> Void in
-            textfield.placeholder = NSLocalizedString("Username", comment: "Username textbox in Authentication prompt")
-            textfield.text = credentials?.user
-        }
-
-        // Add a password textfield.
-        alert.addTextFieldWithConfigurationHandler { (textfield) -> Void in
-            textfield.placeholder = NSLocalizedString("Password", comment: "Password textbox in Authentication prompt")
-            textfield.secureTextEntry = true
-            textfield.text = credentials?.password
-        }
-        
-        viewController.presentViewController(alert, animated: true) { () -> Void in }
-        return deferred
-    }
-    
 }
